+++ src/bookmarks/timeline/TimelineView.tsx
@@ -31,7 +31,7 @@ import { TimelineSkeleton } from './TimelineSkeleton';
 import { VirtualizedTimeline } from './VirtualizedTimeline';
 import { cbService } from '../services/cbService';
 import { bookmarkDB } from '../../utils/bookmarkDB';
-import { getTextSync, getLanguage, clearLanguageCache } from '../utils/i18n';
+import { getTextSync, useI18n } from '../utils/i18n';
 
 /**
  * タイムライン表示コンポーネント
@@ -48,7 +48,9 @@ export function TimelineView() {
     authorUsername: string;
     content: string;
   }>>(new Map());
-  const [languageKey, setLanguageKey] = useState(0); // 言語変更時に再レンダリングをトリガー
+  
+  // i18n機能を一元管理（言語設定の初期化、ストレージ変更の監視、再レンダリングのトリガー）
+  useI18n();
   
   // 編集モードの状態管理
   const [editingName, setEditingName] = useState(false);
@@ -74,41 +76,6 @@ export function TimelineView() {
       setEditingDescription(false);
     }
   }, [selectedCb?.id]);
-
-  // 言語設定を取得してキャッシュに保存
-  useEffect(() => {
-    const initLanguage = async () => {
-      await getLanguage();
-    };
-    initLanguage();
-  }, []);
-
-  // ストレージ変更を監視して言語設定の変更を検知
-  useEffect(() => {
-    const handleStorageChange = (changes: { [key: string]: chrome.storage.StorageChange }, areaName: string) => {
-      if (areaName === 'local' && changes.comiketter_settings) {
-        // 言語設定が変更された場合、キャッシュをクリアして再取得
-        clearLanguageCache();
-        getLanguage().then(() => {
-          // 強制的に再レンダリング（言語キーを変更）
-          setLanguageKey(prev => prev + 1);
-        });
-      }
-    };
-
-    // ストレージ変更イベントをリッスン
-    chrome.storage.onChanged.addListener(handleStorageChange);
-
-    return () => {
-      chrome.storage.onChanged.removeListener(handleStorageChange);
-    };
-  }, []);
-
-  // languageKeyが変更された時に再レンダリングを確実にする
-  useEffect(() => {
-    // languageKeyが変更された時、コンポーネントが再レンダリングされ、
-    // getTextSyncが再実行されるため、新しい言語が反映される
-  }, [languageKey]);
 
   // ストアのshouldEditNameフラグを監視して編集モードに入る
   useEffect(() => {
