+++ src/bookmarks/tweet/Tweet.tsx
@@ -47,7 +47,7 @@ import { useCbStore } from '../state/cbStore';
 import { cbService } from '../services/cbService';
 import { Cb } from '../types/cb';
 import { showNotification } from '@mantine/notifications';
-import { getTextSync, getLanguage, clearLanguageCache } from '../utils/i18n';
+import { getTextSync, useI18n } from '../utils/i18n';
 
 /**
  * 本番用ツイート表示コンポーネント
@@ -78,42 +78,9 @@ export function Tweet({ id, onDelete }: TweetProps) {
   const [initialCbIds, setInitialCbIds] = useState<Set<string>>(new Set());
   const [isLoadingCbs, setIsLoadingCbs] = useState(false);
   const [isSavingCbs, setIsSavingCbs] = useState(false);
-  const [languageKey, setLanguageKey] = useState(0); // 言語変更時に再レンダリングをトリガー
-
-  // 言語設定を取得してキャッシュに保存
-  useEffect(() => {
-    const initLanguage = async () => {
-      await getLanguage();
-    };
-    initLanguage();
-  }, []);
-
-  // ストレージ変更を監視して言語設定の変更を検知
-  useEffect(() => {
-    const handleStorageChange = (changes: { [key: string]: chrome.storage.StorageChange }, areaName: string) => {
-      if (areaName === 'local' && changes.comiketter_settings) {
-        // 言語設定が変更された場合、キャッシュをクリアして再取得
-        clearLanguageCache();
-        getLanguage().then(() => {
-          // 強制的に再レンダリング（言語キーを変更）
-          setLanguageKey(prev => prev + 1);
-        });
-      }
-    };
-
-    // ストレージ変更イベントをリッスン
-    chrome.storage.onChanged.addListener(handleStorageChange);
-
-    return () => {
-      chrome.storage.onChanged.removeListener(handleStorageChange);
-    };
-  }, []);
-
-  // languageKeyが変更された時に再レンダリングを確実にする
-  useEffect(() => {
-    // languageKeyが変更された時、コンポーネントが再レンダリングされ、
-    // getTextSyncが再実行されるため、新しい言語が反映される
-  }, [languageKey]);
+  
+  // i18n機能を一元管理（言語設定の初期化、ストレージ変更の監視、再レンダリングのトリガー）
+  useI18n();
 
   // BookmarkDBからツイート情報を取得
   useEffect(() => {
