+++ src/bookmarks/tweet/TweetEmbedFallback.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Alert, Text, Button, Box } from '@mantine/core';
 import { IconAlertCircle, IconRefresh, IconExternalLink } from '@tabler/icons-react';
 import { formatTweetId } from '../utils/format';
-import { getTextSync, getLanguage, clearLanguageCache } from '../utils/i18n';
+import { getTextSync, useI18n } from '../utils/i18n';
 
 interface TweetEmbedFallbackProps {
   id: string;
@@ -14,42 +14,8 @@ interface TweetEmbedFallbackProps {
  * エラー状態を表示し、再試行やTwitterでの確認を促す
  */
 export function TweetEmbedFallback({ id, onRetry }: TweetEmbedFallbackProps) {
-  const [languageKey, setLanguageKey] = useState(0); // 言語変更時に再レンダリングをトリガー
-
-  // 言語設定を取得してキャッシュに保存
-  useEffect(() => {
-    const initLanguage = async () => {
-      await getLanguage();
-    };
-    initLanguage();
-  }, []);
-
-  // ストレージ変更を監視して言語設定の変更を検知
-  useEffect(() => {
-    const handleStorageChange = (changes: { [key: string]: chrome.storage.StorageChange }, areaName: string) => {
-      if (areaName === 'local' && changes.comiketter_settings) {
-        // 言語設定が変更された場合、キャッシュをクリアして再取得
-        clearLanguageCache();
-        getLanguage().then(() => {
-          // 強制的に再レンダリング（言語キーを変更）
-          setLanguageKey(prev => prev + 1);
-        });
-      }
-    };
-
-    // ストレージ変更イベントをリッスン
-    chrome.storage.onChanged.addListener(handleStorageChange);
-
-    return () => {
-      chrome.storage.onChanged.removeListener(handleStorageChange);
-    };
-  }, []);
-
-  // languageKeyが変更された時に再レンダリングを確実にする
-  useEffect(() => {
-    // languageKeyが変更された時、コンポーネントが再レンダリングされ、
-    // getTextSyncが再実行されるため、新しい言語が反映される
-  }, [languageKey]);
+  // i18n機能を一元管理（言語設定の初期化、ストレージ変更の監視、再レンダリングのトリガー）
+  useI18n();
 
   const handleRetry = () => {
     if (onRetry) {
