+++ src/bookmarks/utils/i18n.ts
@@ -134,3 +134,50 @@ export function useLanguage(): 'ja' | 'en' {
   return language;
 }
 
+/**
+ * React Hook: i18n機能を一元管理
+ * 言語設定の初期化、ストレージ変更の監視、再レンダリングのトリガーを提供
+ * 
+ * @returns 言語変更時に再レンダリングをトリガーするためのキー
+ */
+export function useI18n(): number {
+  const [languageKey, setLanguageKey] = React.useState(0);
+
+  // 言語設定を取得してキャッシュに保存
+  React.useEffect(() => {
+    const initLanguage = async () => {
+      await getLanguage();
+    };
+    initLanguage();
+  }, []);
+
+  // ストレージ変更を監視して言語設定の変更を検知
+  React.useEffect(() => {
+    const handleStorageChange = (changes: { [key: string]: chrome.storage.StorageChange }, areaName: string) => {
+      if (areaName === 'local' && changes.comiketter_settings) {
+        // 言語設定が変更された場合、キャッシュをクリアして再取得
+        clearLanguageCache();
+        getLanguage().then(() => {
+          // 強制的に再レンダリング（言語キーを変更）
+          setLanguageKey(prev => prev + 1);
+        });
+      }
+    };
+
+    // ストレージ変更イベントをリッスン
+    chrome.storage.onChanged.addListener(handleStorageChange);
+
+    return () => {
+      chrome.storage.onChanged.removeListener(handleStorageChange);
+    };
+  }, []);
+
+  // languageKeyが変更された時に再レンダリングを確実にする
+  React.useEffect(() => {
+    // languageKeyが変更された時、コンポーネントが再レンダリングされ、
+    // getTextSyncが再実行されるため、新しい言語が反映される
+  }, [languageKey]);
+
+  return languageKey;
+}
+
